Move DB-independent sequence-DDL guards out of the migration suite
The helper test and static sequence-DDL guards lived in tests/test_migrations.py,
which is marked "migration" and excluded from the coverage unit-test run
(pytest -m "not migration"). As a result build_restart_sequence_sql showed as
uncovered, and the static guards only ran in the DB-gated migration job rather
than on every PR.

Move the four database-independent checks to tests/test_sequence_ddl.py (no
marker, no TEST_DATABASE_URL gate) so they run in the standard suite, count
towards coverage, and guard every PR.
